fix: slicing, reparenting, parsing of "unnamed".
- list-likes slicing
Now provides explicit defaults for slices in case there was a None
in `start`, `end` or `step`. Previous code would treat `stop=0` as
"read till the end of a slice" instead of "read till the first element".
Now None means the former while 0 means the latter.

- parsing of "unnamed"
unnamed declarations are now parsed correctly.

- reparenting an item
Linking same object to 2 parents sequentially would break the previous
parent by swapping the object with None. This None was not checked
in other places which would cause an error.